botegram/requirements.txt
requests
pytelegrambotapi
tritonclient
geventhttpclient
SoundFile
pyyaml